Housekeeping of redundant files
