+++ .gitignore
@@ -1,4 +1,3 @@
-data/downloads/
 package-lock.json
 
 # Logs
