--- bin/seneca-maintain.js
@@ -1,7 +1,0 @@
-#!/usr/bin/env node
-
-const { Maintain } = require('../maintain')
-const maintain = new Maintain()
-
-console.log(maintain.run())
-
