--- data/csv/pluginChecks.csv
@@ -1,23 +1,0 @@
-org/repo,file,url,status,content
-GlenTiki/seneca-load-balancer,README.md,https://github.com/GlenTiki/seneca-load-balancer,FAIL,null
-GlenTiki/seneca-load-balancer,package.json,https://github.com/GlenTiki/seneca-load-balancer,PASS,nonzero_char
-Istar-Eldritch/seneca-sequelize,README.md,https://github.com/Istar-Eldritch/seneca-sequelize,PASS,nonzero_char
-Istar-Eldritch/seneca-sequelize,package.json,https://github.com/Istar-Eldritch/seneca-sequelize,PASS,nonzero_char
-chico/seneca-perm-editor,README.md,https://github.com/chico/seneca-perm-editor,PASS,nonzero_char
-chico/seneca-perm-editor,package.json,https://github.com/chico/seneca-perm-editor,PASS,nonzero_char
-rjrodger/seneca-todo-flex,README.md,https://github.com/rjrodger/seneca-todo-flex,PASS,nonzero_char
-rjrodger/seneca-todo-flex,package.json,https://github.com/rjrodger/seneca-todo-flex,PASS,nonzero_char
-salmanm/seneca-calendar-flex,README.md,https://github.com/salmanm/seneca-calendar-flex,FAIL,null
-salmanm/seneca-calendar-flex,package.json,https://github.com/salmanm/seneca-calendar-flex,PASS,nonzero_char
-senecajs-attic/seneca-pubsub,README.md,https://github.com/senecajs-attic/seneca-pubsub,PASS,nonzero_char
-senecajs-attic/seneca-pubsub,package.json,https://github.com/senecajs-attic/seneca-pubsub,PASS,nonzero_char
-senecajs-labs/seneca-import-csv,README.md,https://github.com/senecajs-labs/seneca-import-csv,PASS,nonzero_char
-senecajs-labs/seneca-import-csv,package.json,https://github.com/senecajs-labs/seneca-import-csv,PASS,nonzero_char
-senecajs/seneca-parambulator,README.md,https://github.com/senecajs/seneca-parambulator,PASS,nonzero_char
-senecajs/seneca-parambulator,package.json,https://github.com/senecajs/seneca-parambulator,PASS,nonzero_char
-senecajs/seneca-transport,README.md,https://github.com/senecajs/seneca-transport,PASS,nonzero_char
-senecajs/seneca-transport,package.json,https://github.com/senecajs/seneca-transport,PASS,nonzero_char
-timugz/seneca-token,README.md,https://github.com/timugz/seneca-token,FAIL,null
-timugz/seneca-token,package.json,https://github.com/timugz/seneca-token,PASS,nonzero_char
-voxgig/seneca-promisify,README.md,https://github.com/voxgig/seneca-promisify,PASS,nonzero_char
-voxgig/seneca-promisify,package.json,https://github.com/voxgig/seneca-promisify,PASS,nonzero_char
