--- data/json/pluginChecks.json
@@ -1,1 +1,0 @@
-[{"org/repo":"GlenTiki/seneca-load-balancer","file":"README.md","url":"https://github.com/GlenTiki/seneca-load-balancer","status":"FAIL","content":"null"},{"org/repo":"GlenTiki/seneca-load-balancer","file":"package.json","url":"https://github.com/GlenTiki/seneca-load-balancer","status":"PASS","content":"nonzero_char"},{"org/repo":"Istar-Eldritch/seneca-sequelize","file":"README.md","url":"https://github.com/Istar-Eldritch/seneca-sequelize","status":"PASS","content":"nonzero_char"},{"org/repo":"Istar-Eldritch/seneca-sequelize","file":"package.json","url":"https://github.com/Istar-Eldritch/seneca-sequelize","status":"PASS","content":"nonzero_char"},{"org/repo":"chico/seneca-perm-editor","file":"README.md","url":"https://github.com/chico/seneca-perm-editor","status":"PASS","content":"nonzero_char"},{"org/repo":"chico/seneca-perm-editor","file":"package.json","url":"https://github.com/chico/seneca-perm-editor","status":"PASS","content":"nonzero_char"},{"org/repo":"rjrodger/seneca-todo-flex","file":"README.md","url":"https://github.com/rjrodger/seneca-todo-flex","status":"PASS","content":"nonzero_char"},{"org/repo":"rjrodger/seneca-todo-flex","file":"package.json","url":"https://github.com/rjrodger/seneca-todo-flex","status":"PASS","content":"nonzero_char"},{"org/repo":"salmanm/seneca-calendar-flex","file":"README.md","url":"https://github.com/salmanm/seneca-calendar-flex","status":"FAIL","content":"null"},{"org/repo":"salmanm/seneca-calendar-flex","file":"package.json","url":"https://github.com/salmanm/seneca-calendar-flex","status":"PASS","content":"nonzero_char"},{"org/repo":"senecajs-attic/seneca-pubsub","file":"README.md","url":"https://github.com/senecajs-attic/seneca-pubsub","status":"PASS","content":"nonzero_char"},{"org/repo":"senecajs-attic/seneca-pubsub","file":"package.json","url":"https://github.com/senecajs-attic/seneca-pubsub","status":"PASS","content":"nonzero_char"},{"org/repo":"senecajs-labs/seneca-import-csv","file":"README.md","url":"https://github.com/senecajs-labs/seneca-import-csv","status":"PASS","content":"nonzero_char"},{"org/repo":"senecajs-labs/seneca-import-csv","file":"package.json","url":"https://github.com/senecajs-labs/seneca-import-csv","status":"PASS","content":"nonzero_char"},{"org/repo":"senecajs/seneca-parambulator","file":"README.md","url":"https://github.com/senecajs/seneca-parambulator","status":"PASS","content":"nonzero_char"},{"org/repo":"senecajs/seneca-parambulator","file":"package.json","url":"https://github.com/senecajs/seneca-parambulator","status":"PASS","content":"nonzero_char"},{"org/repo":"senecajs/seneca-transport","file":"README.md","url":"https://github.com/senecajs/seneca-transport","status":"PASS","content":"nonzero_char"},{"org/repo":"senecajs/seneca-transport","file":"package.json","url":"https://github.com/senecajs/seneca-transport","status":"PASS","content":"nonzero_char"},{"org/repo":"timugz/seneca-token","file":"README.md","url":"https://github.com/timugz/seneca-token","status":"FAIL","content":"null"},{"org/repo":"timugz/seneca-token","file":"package.json","url":"https://github.com/timugz/seneca-token","status":"PASS","content":"nonzero_char"},{"org/repo":"voxgig/seneca-promisify","file":"README.md","url":"https://github.com/voxgig/seneca-promisify","status":"PASS","content":"nonzero_char"},{"org/repo":"voxgig/seneca-promisify","file":"package.json","url":"https://github.com/voxgig/seneca-promisify","status":"PASS","content":"nonzero_char"}][{"orgRepo":"GlenTiki/seneca-load-balancer","file":"README.md","url":"https://github.com/GlenTiki/seneca-load-balancer","status":"FAIL","content":"null"},{"orgRepo":"GlenTiki/seneca-load-balancer","file":"package.json","url":"https://github.com/GlenTiki/seneca-load-balancer","status":"PASS","content":"nonzero_char"},{"orgRepo":"Istar-Eldritch/seneca-sequelize","file":"README.md","url":"https://github.com/Istar-Eldritch/seneca-sequelize","status":"PASS","content":"nonzero_char"},{"orgRepo":"Istar-Eldritch/seneca-sequelize","file":"package.json","url":"https://github.com/Istar-Eldritch/seneca-sequelize","status":"PASS","content":"nonzero_char"},{"orgRepo":"chico/seneca-perm-editor","file":"README.md","url":"https://github.com/chico/seneca-perm-editor","status":"PASS","content":"nonzero_char"},{"orgRepo":"chico/seneca-perm-editor","file":"package.json","url":"https://github.com/chico/seneca-perm-editor","status":"PASS","content":"nonzero_char"},{"orgRepo":"rjrodger/seneca-todo-flex","file":"README.md","url":"https://github.com/rjrodger/seneca-todo-flex","status":"PASS","content":"nonzero_char"},{"orgRepo":"rjrodger/seneca-todo-flex","file":"package.json","url":"https://github.com/rjrodger/seneca-todo-flex","status":"PASS","content":"nonzero_char"},{"orgRepo":"salmanm/seneca-calendar-flex","file":"README.md","url":"https://github.com/salmanm/seneca-calendar-flex","status":"FAIL","content":"null"},{"orgRepo":"salmanm/seneca-calendar-flex","file":"package.json","url":"https://github.com/salmanm/seneca-calendar-flex","status":"PASS","content":"nonzero_char"},{"orgRepo":"senecajs-attic/seneca-pubsub","file":"README.md","url":"https://github.com/senecajs-attic/seneca-pubsub","status":"PASS","content":"nonzero_char"},{"orgRepo":"senecajs-attic/seneca-pubsub","file":"package.json","url":"https://github.com/senecajs-attic/seneca-pubsub","status":"PASS","content":"nonzero_char"},{"orgRepo":"senecajs-labs/seneca-import-csv","file":"README.md","url":"https://github.com/senecajs-labs/seneca-import-csv","status":"PASS","content":"nonzero_char"},{"orgRepo":"senecajs-labs/seneca-import-csv","file":"package.json","url":"https://github.com/senecajs-labs/seneca-import-csv","status":"PASS","content":"nonzero_char"},{"orgRepo":"senecajs/seneca-parambulator","file":"README.md","url":"https://github.com/senecajs/seneca-parambulator","status":"PASS","content":"nonzero_char"},{"orgRepo":"senecajs/seneca-parambulator","file":"package.json","url":"https://github.com/senecajs/seneca-parambulator","status":"PASS","content":"nonzero_char"},{"orgRepo":"senecajs/seneca-transport","file":"README.md","url":"https://github.com/senecajs/seneca-transport","status":"PASS","content":"nonzero_char"},{"orgRepo":"senecajs/seneca-transport","file":"package.json","url":"https://github.com/senecajs/seneca-transport","status":"PASS","content":"nonzero_char"},{"orgRepo":"timugz/seneca-token","file":"README.md","url":"https://github.com/timugz/seneca-token","status":"FAIL","content":"null"},{"orgRepo":"timugz/seneca-token","file":"package.json","url":"https://github.com/timugz/seneca-token","status":"PASS","content":"nonzero_char"},{"orgRepo":"voxgig/seneca-promisify","file":"README.md","url":"https://github.com/voxgig/seneca-promisify","status":"PASS","content":"nonzero_char"},{"orgRepo":"voxgig/seneca-promisify","file":"package.json","url":"https://github.com/voxgig/seneca-promisify","status":"PASS","content":"nonzero_char"}]
