--- data/json/sheetsCredentials.json
@@ -1,1 +1,0 @@
-{"installed":{"client_id":"790973298961-cq9k9tj8puuofm8alfg939p8ph36tbnm.apps.googleusercontent.com","project_id":"senecajs-repo-maintain","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://oauth2.googleapis.com/token","auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs","client_secret":"NhYev7d9BIzUPEKngbNjzJUX","redirect_uris":["urn:ietf:wg:oauth:2.0:oob","http://localhost"]}}
