--- data/json/token.json
@@ -1,1 +1,0 @@
-{"access_token":"ya29.a0ARrdaM_2-_43TA4kIoNdMwxtCSLOMFnksYzZsROKwlfobTa5WVYXU6xOCqxso4rendTTl-weMXvrd26Z6K1EMrupeGRoKkABtDcfS0G8-tf1dfGqqEq3CHD1ysq4a3u9zl52MNruMJrq7J-ydNeZ-2EW6q36","refresh_token":"1//07vCRwdB7sKH2CgYIARAAGAcSNwF-L9IrjKuPFkPC6wugDMOUK_CdNTMzDsctGg9deGift6w0_3KdGeHAMczd6VvACE_7UINbq98","scope":"https://www.googleapis.com/auth/spreadsheets.readonly","token_type":"Bearer","expiry_date":1625058826100}
