--- data/txt/2lines.txt
@@ -1,2 +1,0 @@
-alpha
-beta
