--- design/checks/NOTES.md
@@ -1,1 +1,0 @@
-* checks.js is the declarative definition of the checks to run
