--- design/checks/results.csv
@@ -1,10 +1,0 @@
-Name,                   Pass,   exist_readme,   exist_pkgjson,  exist_license,  content_readme
-@senecajs/foo,          2,      P,              P,              F,              F
-seneca-tranport,        3,      P,              P,              F,              P              
-...
-
-
-
-
-
-
