--- script/isolate.js
@@ -1,10 +1,0 @@
-const Fs = require('fs')
-const _ = require('underscore')
-
-const filter = require('../data/json/filter.json')
-
-async function doIsolate() {
-  console.log("ITEMS : ", filter.length)
-}
-
-doIsolate()
