--- script/plugins.js
@@ -1,117 +1,0 @@
-const Fs = require('fs')
-
-plugins = {
-  'senecajs/seneca-transport': { kind: 'plugin' },
-  'Istar-Eldritch/seneca-sequelize': { kind: 'plugin' },
-  'senecajs-attic/seneca-pubsub': { kind: 'plugin' },
-  'chico/seneca-perm-editor': { kind: 'plugin' },
-  'timugz/seneca-token': { kind: 'plugin' },
-  'voxgig/seneca-promisify': { kind: 'plugin' },
-  'senecajs/seneca-parambulator': { kind: 'plugin' },
-  'rjrodger/seneca-todo-flex': { kind: 'plugin' },
-  'senecajs-labs/seneca-import-csv': { kind: 'plugin' },
-  'GlenTiki/seneca-load-balancer': { kind: 'plugin' },
-  'senecajs/seneca-entity': { kind: 'plugin' },
-  'senecajs/seneca-log-filter': { kind: 'plugin' },
-  'eoinsha/node-seneca-base': { kind: 'plugin' },
-  'senecajs/seneca-parambulator': { kind: 'plugin' },
-  'senecajs-labs/seneca-store-query': { kind: 'plugin' },
-  'voxgig/seneca-populate': { kind: 'plugin' },
-  'voxgig/seneca-allow': { kind: 'plugin' },
-  'voxgig/seneca-owner': { kind: 'plugin' },
-  'wyvernzora/seneca-as-promised': { kind: 'plugin' },
-  'voxgig/seneca-member': { kind: 'plugin' },
-  'voxgig/seneca-promisify': { kind: 'plugin' },
-  'senecajs/seneca-apikey': { kind: 'plugin' },
-  'nscale/seneca-balance-client': { kind: 'plugin' },
-  'senecajs/seneca-mem-store': { kind: 'plugin' },
-  'senecajs/seneca-postgres-store': { kind: 'plugin' },
-  'senecajs/seneca-vote': { kind: 'plugin' },
-  'senecajs/seneca-mail': { kind: 'plugin' },
-  'senecajs/seneca-mongo-store': { kind: 'plugin' },
-  'senecajs/seneca-dynamo-store': { kind: 'plugin' },
-  'senecajs/seneca-basic': { kind: 'plugin' },
-  'senecajs/seneca-repl': { kind: 'plugin' },
-  'senecajs/seneca-graph': { kind: 'plugin' },
-  'senecajs/seneca-joi': { kind: 'plugin' },
-  'senecajs/seneca-memcached-cache': { kind: 'plugin' },
-  'voxgig/seneca-group': { kind: 'plugin' },
-  'voxgig/seneca-sendgrid-mail': { kind: 'plugin' },
-  'voxgig/seneca-kv': { kind: 'plugin' },
-  'voxgig/seneca-redis-kv': { kind: 'plugin' },
-  'senecajs/seneca-entity': { kind: 'plugin' },
-  'senecajs/seneca-amqp-transport': { kind: 'plugin' },
-  'senecajs/seneca-test-plugin': { kind: 'plugin' },
-  'senecajs/seneca-level-store': { kind: 'plugin' },
-  'senecajs/seneca-auth': { kind: 'plugin' },
-  'salmanm/seneca-calendar-flex': { kind: 'plugin' },
-  'senecajs/seneca-transport': { kind: 'plugin' },
-  'senecajs/seneca-parambulator': { kind: 'plugin' },
-  'senecajs/seneca-entity': { kind: 'plugin' },
-  'senecajs/seneca-log-filter': { kind: 'plugin' },
-  'senecajs/seneca-apikey': { kind: 'plugin' },
-  'senecajs/seneca-mem-store': { kind: 'plugin' },
-  'senecajs/seneca-postgres-store': { kind: 'plugin' },
-  'senecajs/seneca-vote': { kind: 'plugin' },
-  'senecajs/seneca-mail': { kind: 'plugin' },
-  'senecajs/seneca-mongo-store': { kind: 'plugin' },
-  'senecajs/seneca-dynamo-store': { kind: 'plugin' },
-  'senecajs/seneca-basic': { kind: 'plugin' },
-  'senecajs/seneca-repl': { kind: 'plugin' },
-  'senecajs/seneca-graph': { kind: 'plugin' },
-  'senecajs/seneca-joi': { kind: 'plugin' },
-  'senecajs/seneca-memcached-cache': { kind: 'plugin' },
-  'senecajs/seneca-amqp-transport': { kind: 'plugin' },
-  'senecajs/seneca-test-plugin': { kind: 'plugin' },
-  'senecajs/seneca-level-store': { kind: 'plugin' },
-  'senecajs/seneca-auth': { kind: 'plugin' },
-  'senecajs/seneca-cache': { kind: 'plugin' },
-  'senecajs/seneca-user': { kind: 'plugin' },
-  'senecajs/seneca-mesh': { kind: 'plugin' },
-  'senecajs/seneca-balance-client': { kind: 'plugin' },
-  'senecajs/seneca-redis-cache': { kind: 'plugin' },
-  'senecajs/seneca-entity-cache': { kind: 'plugin' },
-  'senecajs/seneca-cache-test': { kind: 'plugin' },
-  'senecajs/seneca-redis-pubsub-transport': { kind: 'plugin' },
-  'senecajs/seneca-redis-queue-transport': { kind: 'plugin' },
-  'senecajs/seneca-registry': { kind: 'plugin' },
-  'senecajs/seneca-jsonfile-store': { kind: 'plugin' },
-  'senecajs/seneca-pino-logger': { kind: 'plugin' },
-  'senecajs/seneca-consul-registry': { kind: 'plugin' },
-  'senecajs/seneca-redis-store': { kind: 'plugin' },
-  'senecajs/seneca-logstash-logger': { kind: 'plugin' },
-  'senecajs/seneca-logentries-logger': { kind: 'plugin' },
-  'voxgig/seneca-populate': { kind: 'plugin' },
-  'voxgig/seneca-allow': { kind: 'plugin' },
-  'voxgig/seneca-owner': { kind: 'plugin' },
-  'voxgig/seneca-member': { kind: 'plugin' },
-  'voxgig/seneca-promisify': { kind: 'plugin' },
-  'voxgig/seneca-group': { kind: 'plugin' },
-  'voxgig/seneca-kv': { kind: 'plugin' },
-  'voxgig/seneca-redis-kv': { kind: 'plugin' },
-  'voxgig/seneca-debug': { kind: 'plugin' },
-  'voxgig/seneca-msg-run': { kind: 'plugin' },
-  'voxgig/seneca-hapi': { kind: 'plugin' },
-  'voxgig/seneca-user-telemetry': { kind: 'plugin' },
-  'voxgig/seneca-doc': { kind: 'plugin' },
-  'senecajs-labs/seneca-sqlite-store': { kind: 'plugin' },
-  'senecajs-labs/seneca-zipkin-tracer': { kind: 'plugin' },
-  'senecajs-labs/seneca-dynamo-store': { kind: 'plugin' },
-  'senecajs-labs/seneca-hana-store': { kind: 'plugin' },
-  'senecajs-labs/seneca-shard-store': { kind: 'plugin' },
-  'senecajs-labs/seneca-store-query': { kind: 'plugin' },
-  'senecajs-labs/seneca-standard-query': { kind: 'plugin' },
-  'senecajs-labs/seneca-cluster': { kind: 'plugin' },
-  'senecajs-labs/seneca-queue': { kind: 'plugin' },
-  'senecajs-labs/seneca-quartz-scheduler': { kind: 'plugin' },
-  'senecajs-labs/seneca-cassandra-store': { kind: 'plugin' },
-  'senecajs-labs/seneca-import-csv': { kind: 'plugin' },
-  'senecajs-labs/seneca-entity-save-stream': { kind: 'plugin' },
-  'senecajs-labs/seneca-elasticsearch': { kind: 'plugin' },
-  'senecajs-labs/seneca-act-stream': { kind: 'plugin' },
-  'senecajs-labs/seneca-rabbitmq-transport': { kind: 'plugin' },
-}
-
-json = JSON.stringify(plugins)
-Fs.writeFileSync('../data/json/plugins.json', json)
-console.log('Plugin list formatted to JSON. See plugins.json for data.')
