--- script/tests/token.json
@@ -1,1 +1,0 @@
-{"access_token":"ya29.a0AfH6SMB-vRBmhJ6h9xHqUfjr56rGqqq7yWWLeeCurrvbYK13YElPMfYof10ZyiAwjMN88yR_29vSxG1vp0-MOwrt3unWsE0P6PRK-zajuUoNydE41ltLa8adF12bFGuKg4635aPz4wHqieLl5X3iXSdXfJfU","refresh_token":"1//03qNRH1D3hDduCgYIARAAGAMSNwF-L9IrxqJiu-rCUzTgHHzY6lis9MrpMWAHzyNK9l5shkmSq3UraTqvgRLbMKhqkzBNrYqJUoc","scope":"https://www.googleapis.com/auth/spreadsheets.readonly","token_type":"Bearer","expiry_date":1623779062767}
